pyproject.toml
[tool.poetry]
name = "freedesktop-icons"
version = "1.0.0a1"
description = "Find icon paths according to the freedesktop icon theme specification."
authors = ["Ash Berlin-Taylor <ash_github@firemirror.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]
vulture = "^1.0"
bandit = "^1.6"
safety = "^1.8"
isort = "^5.3"
flake8-bugbear = "^19.8"
black = {version = "*", allow-prereleases = true}
mypy = "~0.812"
ipython = "^7.7"
pytest = "^6.0"
pytest-cov = "^2.7"
pytest-deadfixtures = "^2.2.1"
pep8-naming = "^0.8.2"
cruft = "^2.2"
furo = "^2021.6.18-beta.36"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 160
skip-string-normalization = true

[tool.isort]
balanced_wrapping = true
default_section = "THIRDPARTY"
include_trailing_comma = true
known_first_party = [ "freedesktop_icons", "tests" ]
line_length = 160
multi_line_output = 3

[tool.cruft]
skip = [
  "freedesktop_icons/__init__.py",
  "tests",
  "docs/",
  "pyproject.toml",
  "poetry.lock",
  "CHANGELOG.md",
  "README.md",
  ".flake8",
  ".git",
]

[tool.pytest.ini_options]
addopts = "--strict-markers --doctest-modules"

[tool.coverage.run]
omit = ["tests/*","**/__main__.py"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
]
